Refactor for codesense
